Reapply "Merge remote-tracking branch 'origin/fix/issue-271-float32-value-bus' into integration/0.3.0-entity"
This reverts commit 8d277b5ca25f0fd31cd63ae7b545de76406e9cf8.

+++ internal/agent/probes/memory/memoryProbe_windows.go
@@ -207,7 +207,7 @@ func (w *windowsMemoryCollector) Collect(timestamp time.Time) ([]data_store.Data
 		dataPoints = append(dataPoints, data_store.DataPoint{
 			Name:      name,
 			Timestamp: timestamp,
-			Value:     float32(value),
+			Value:     value,
 			Tags:      metricTags,
 		})
 
@@ -221,7 +221,7 @@ func (w *windowsMemoryCollector) Collect(timestamp time.Time) ([]data_store.Data
 			dataPoints = append(dataPoints, data_store.DataPoint{
 				Name:      "memory_used_percent",
 				Timestamp: timestamp,
-				Value:     float32(usedPercent),
+				Value:     usedPercent,
 				Tags:      baseTags,
 			})
 		}
